Fix grammar errors in folders. migrate forms to Forms folder. Migrate home schema to Schemas folder. Migrate homeForm utils to Utils folder
